+++ src/routes/api.ts
@@ -67,6 +67,7 @@ const routes = new Hono()
 // Auth
 routes.post("/auth/register", zValidator("json", RegisterValidator, validationHook), (c) => authController.register(c))
 routes.post("/auth/login", zValidator("json", LoginValidator, validationHook), (c) => authController.login(c))
+routes.post("/auth/nusawork-login", zValidator("json", LoginValidator, validationHook), (c) => authController.nusaworkLogin(c))
 routes.post("/auth/google", zValidator("json", GoogleLoginValidator, validationHook), (c) => authController.google(c))
 routes.post("/auth/forgot-password", zValidator("json", ForgotPasswordValidator, validationHook), (c) => authController.forgotPassword(c))
 routes.get("/auth/validate-reset-token", (c) => authController.validateResetToken(c))